+++ src/app/item-review/item-review.component.js
@@ -1,3 +1,4 @@
+import _ from 'underscore';
 import template from './item-review.html';
 import './item-review.scss';
 
@@ -10,6 +11,7 @@ function ItemReviewController(dimSettingsService, dimDestinyTrackerService, $sco
   vm.submitted = false;
   vm.hasUserReview = vm.item.userRating;
   vm.expandReview = vm.item.isLocallyCached;
+  vm.toggledFlags = [];
 
   vm.procon = false; // TODO: turn this back on..
   vm.aggregate = {
@@ -25,7 +27,44 @@ function ItemReviewController(dimSettingsService, dimDestinyTrackerService, $sco
     vm.expandReview = !vm.expandReview;
   };
 
-  vm.editReview = function() {
+  vm.clickReview = function(reviewId) {
+    const review = _.find(vm.item.writtenReviews, { reviewId: reviewId });
+
+    if (review.isReviewer) {
+      vm.editReview();
+    }
+    else if (!review.isHighlighted) {
+      vm.openFlagContext(reviewId);
+    }
+  };
+
+  vm.openFlagContext = function(reviewId) {
+    const review = _.find(vm.item.writtenReviews, { reviewId: reviewId });
+
+    if ((review.isReviewer) || (review.isHighlighted)) {
+      return;
+    }
+
+    const toggledReviewIndex = vm.toggledFlags.indexOf(reviewId);
+
+    if (toggledReviewIndex === -1) {
+      vm.toggledFlags.push(reviewId);
+    }
+  };
+
+  vm.closeFlagContext = function(reviewId) {
+    const toggledReviewIndex = vm.toggledFlags.indexOf(reviewId);
+
+    vm.toggledFlags.splice(toggledReviewIndex);
+  };
+
+  vm.editReview = function(reviewId) {
+    const review = _.find(vm.item.writtenReviews, { reviewId: reviewId });
+
+    if (!review.isReviewer) {
+      return;
+    }
+
     vm.expandReview = true;
   };
 
@@ -48,6 +87,12 @@ function ItemReviewController(dimSettingsService, dimDestinyTrackerService, $sco
 
     dimDestinyTrackerService.updateCachedUserRankings(item,
                                                       userReview);
+  };
+
+  vm.reportReview = function(reviewId) {
+    const review = _.find(vm.item.writtenReviews, { reviewId: reviewId });
+
+    dimDestinyTrackerService.reportReview(review);
   };
 
   vm.toUserReview = function(item) {