+++ src/app/loadout-builder/loadout-builder.component.js
@@ -4,12 +4,15 @@ import template from './loadout-builder.html';
 import intellectIcon from 'app/images/intellect.png';
 import disciplineIcon from 'app/images/discipline.png';
 import strengthIcon from 'app/images/strength.png';
-import { getBonus } from '../services/store/character-utils';
+import { getBonus } from '../inventory/store/character-utils';
 
 export const LoadoutBuilderComponent = {
   controller: LoadoutBuilderController,
-  template: template,
-  controllerAs: 'vm'
+  template,
+  controllerAs: 'vm',
+  bindings: {
+    account: '<'
+  }
 };
 
 function LoadoutBuilderController($scope, $state, $q, $timeout, $i18next, dimSettingsService, dimStoreService, ngDialog, dimLoadoutService, dimDefinitions, dimVendorService) {
@@ -323,6 +326,7 @@ function LoadoutBuilderController($scope, $state, $q, $timeout, $i18next, dimSet
         vm.lockeditems = { Helmet: null, Gauntlets: null, Chest: null, Leg: null, ClassItem: null, Artifact: null, Ghost: null };
         vm.lockedperks = { Helmet: {}, Gauntlets: {}, Chest: {}, Leg: {}, ClassItem: {}, Artifact: {}, Ghost: {} };
         vm.excludeditems = _.filter(vm.excludeditems, (item) => { return item.hash === 2672107540; });
+        vm.activesets = null;
         vm.highestsets = vm.getSetBucketsStep(vm.active);
       },
       onActiveSetsChange: function() {
@@ -444,6 +448,7 @@ function LoadoutBuilderController($scope, $state, $q, $timeout, $i18next, dimSet
       },
       clearLocked: function() {
         vm.lockeditems = { Helmet: null, Gauntlets: null, Chest: null, Leg: null, ClassItem: null, Artifact: null, Ghost: null };
+        vm.activesets = null;
         vm.highestsets = vm.getSetBucketsStep(vm.active);
         if (vm.progress < 1.0) {
           vm.lockedchanged = true;
@@ -616,7 +621,9 @@ function LoadoutBuilderController($scope, $state, $q, $timeout, $i18next, dimSet
             }
           }
 
-          vm.activesets = vm.allSetTiers[1];
+          if (!vm.allSetTiers.includes(vm.activesets)) {
+            vm.activesets = vm.allSetTiers[1];
+          }
           vm.activeHighestSets = getActiveHighestSets(setMap, vm.activesets);
           vm.collapsedConfigs = [false, false, false, false, false, false, false, false, false, false];
 
@@ -634,6 +641,7 @@ function LoadoutBuilderController($scope, $state, $q, $timeout, $i18next, dimSet
       getStore: dimStoreService.getStore,
       getItems: function() {
         const stores = dimStoreService.getStores();
+        vm.stores = stores;
 
         if (stores.length === 0) {
           $state.go('destiny1.inventory');